[#39] fix: key값 없는 오류 수정, date잘못 사용 오류 수정
